backend/requirements.txt
Flask==3.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
PyYAML==6.0
python-dotenv==1.0.0
Werkzeug==3.0.0
opencv-python==4.8.1.78
